migrate gfiPie to V3
