--- gfiPie/index.js
@@ -1,5 +1,0 @@
-import component from "./components/GfiPie.vue";
-
-export default {
-    component: component
-};
