+++ masterportalVersion2/gfiThemes/gfiPie/index.js
@@ -1,0 +1,5 @@
+import component from "./components/GfiPie.vue";
+
+export default {
+    component: component
+};
