+++ masterportalVersion2/wiengfi/store/gettersWienGfi.js
