+++ masterportalVersion3/bulkDownload/store/gettersBulkDownload.js
