+++ masterportalVersion3/bulkDownload/store/indexBulkDownload.js
