+++ masterportalVersion3/bulkDownload/store/stateBulkDownload.js
