+++ masterportalVersion3/wfsSumQuery/components/WfsSumQuery.vue
