+++ masterportalVersion3/wfsSumQuery/doc/config.json.de.md
