+++ masterportalVersion3/wfsSumQuery/doc/config.json.md
