+++ masterportalVersion3/wfsSumQuery/index.js
