+++ masterportalVersion3/wfsSumQuery/locales/de/additional.json
