+++ masterportalVersion3/wfsSumQuery/store/actionsWfsSumQuery.js
