+++ masterportalVersion3/wfsSumQuery/store/gettersWfsSumQuery.js
