+++ masterportalVersion3/wfsSumQuery/store/indexWfsSumQuery.js
