+++ masterportalVersion3/wfsSumQuery/store/stateWfsSumQuery.js
